pyproject.toml
[tool.poetry]
name = "fabrik-terraform-poc"
version = "0.2.0"
description = "Python wrapper for terragrunt"
readme = "README.md"
authors = ["Lee Myring <29373851+thinkstack@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-terraform = "^0.10.1"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
bandit = "^1.7.5"
black = "^23.3.0"
pre-commit = "^3.3.2"
pytest = "^7.3.1"
pytest-black = "^0.3.12"
pytest-cov = "^4.0.0"
pytest-pycodestyle = "^2.3.1"
python-semantic-release = "^7.33.5"
reorder-python-imports = "^3.9.0"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.semantic_release]
branch = "main"
upload_to_pypi = false
upload_to_release = false
upload_to_repository = false
version_variable = [
    "pyproject.toml:version",
    'setup.py:__version__',
]
